Required changes

+++ authorize.go
@@ -35,7 +35,7 @@ type AuthorizeRequest struct {
 
 	// Token expiration in seconds. Change if different from default.
 	// If type = TOKEN, this expiration will be for the ACCESS token.
-	Expiration int32
+	Expiration int64
 
 	// Data to be passed to storage. Not used by the library.
 	UserData interface{}
@@ -58,7 +58,7 @@ type AuthorizeData struct {
 	Code string
 
 	// Token expiration in seconds
-	ExpiresIn int32
+	ExpiresIn int64
 
 	// Requested scope
 	Scope string
@@ -153,7 +153,7 @@ func (s *Server) HandleAuthorizeRequest(w *Response, r *http.Request) *Authorize
 		w.InternalError = err
 		return nil
 	} else {
-		ret.RedirectUri =  realRedirectUri
+		ret.RedirectUri = realRedirectUri
 	}
 
 	w.SetRedirect(ret.RedirectUri)
